test: set up pytest (#27)
* refactor: move Python code into a package
* test: set up pytest
* ci: set up testing on GitHub Actions
* docs: add test coverage badge
* test: create test_output.py
* test: create test_explore_lines.py

+++ pathfinder/explore_lines.py
@@ -8,7 +8,7 @@ def get_explore_lines(search_area_lines):
     :returns: Number of lines to explore either side of the search area.
     """
     # Get setting values from Vim variables
-    explore_scale = int(vim.vars["pf_explore_scale"])
+    explore_scale = float(vim.vars["pf_explore_scale"])
     max_explore = int(vim.vars["pf_max_explore"])
 
     if explore_scale < 0:
